打磨三端：clippy 清零 + C ABI(flexui-ffi) 支持 Linux
清理全工作区 clippy 告警至 0（mac/linux 均 0 警告）：
- core/dispatch.rs: set_property 的 match 去掉逐臂 return，改尾表达式返回。
- core/scroll.rs: ScrollBarVisibility 用 #[derive(Default)]+#[default]。
- core/widget.rs: WidgetProperty 加 large_enum_variant allow（一次性载体，装箱无益）并注明。
- native-menu: popup_for_nsview/hwnd 补 # Safety 文档。
- macos/view.rs: 复杂类型抽出 DeferredNativeCallback 别名。
- linux/canvas.rs: 去掉 rect.left()/top() 冗余 as f32；window.rs: 用 RangeInclusive::contains。
- gallery: 派生 Default、is_multiple_of、显式忽略 read 返回值。

C ABI 打通 Linux（此前仅 mac/win，Linux 返回 -100 无后端）：
- flexui-linux 新增单窗口 run()（包一层 run_multi），与 mac 后端对位。
- flexui-ffi 增加 target=linux 后端选择与依赖；11 处 backend 门控 + 占位 stub
  均纳入 linux；flex_run/clipboard/dialog 等 C 入口在 Linux 上可用。
- Linux 上 libflexui_ffi.so 正常产出，flexui-ffi 单测通过。

三端验证：mac clippy 0 警告 +141 测试通过；linux clippy 0 警告 + ffi 测试通过；
windows 交叉编译通过。

+++ crates/flexui-linux/src/canvas.rs
@@ -502,8 +502,8 @@ impl Canvas for CairoCanvas {
                 self.blit_sub(&surface, (0.0, 0.0, sw, sh), rect, tint);
             }
             ImageFit::Center => {
-                let x = rect.left() as f32 + (rect.size.width - lw as f32) / 2.0;
-                let y = rect.top() as f32 + (rect.size.height - lh as f32) / 2.0;
+                let x = rect.left() + (rect.size.width - lw as f32) / 2.0;
+                let y = rect.top() + (rect.size.height - lh as f32) / 2.0;
                 self.blit_sub(&surface, (0.0, 0.0, sw, sh), Rect::new(x, y, lw as f32, lh as f32), tint);
             }
             ImageFit::Tile => {
